requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
environ==1.0
idna==3.10
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.3
sslcommerz-lib==1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.8.2
